update developer dashboard
