+++ client/src/App.js
@@ -7,14 +7,13 @@ import AnalyticsController from "./components/Shared/AnalyticsController";
 import ScrollToTopControlller from "./components/Shared/ScrollToTopControlller";
 import PageTitleController from "./components/Shared/PageTitleController";
 import Notifications from "./components/Notification/Notifications";
-import Overlay from "./components/Shared/Overlay";
-import SaveDashboardModal from "./components/Dashboards/SaveDashboardModal";
+import SaveDashboardModal from "./components/SaveDashboardModal/SaveDashboardModal";
 import Login from "./components/Login";
 import Main from "./components/Main";
 
 // import "./styles/main.scss";
 
-const App = ({ initUser, overlay, saveDashboardModal }) => {
+const App = ({ initUser, saveDashboardModal }) => {
   useEffect(() => {
     initUser();
     // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -26,7 +25,6 @@ const App = ({ initUser, overlay, saveDashboardModal }) => {
         <PageTitleController />
         <ScrollToTopControlller />
         <Notifications />
-        {overlay && <Overlay />}
         {saveDashboardModal && <SaveDashboardModal />}
         <Switch>
           <Route exact path="/login" render={() => <Login />} />
@@ -41,7 +39,6 @@ const App = ({ initUser, overlay, saveDashboardModal }) => {
 
 const mapStateToProps = (state) => {
   return {
-    overlay: state.navigation.overlay,
     saveDashboardModal: state.navigation.modal[CONSTANTS.MODALS.SAVEDASHBOARD],
   };
 };
