+++ client/src/components/Dashboards/DeveloperDashboard.js
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import { connect } from "react-redux";
 import Tooltip from "../Shared/Tooltip";
+import Button from "../Shared/Button";
 import {
   cumulativeChartParse,
   annualChartParse,
@@ -15,9 +16,6 @@ import {
 } from "../../utils/dashboardHelper";
 import { developerTooltips } from "../../static/tooltips";
 import { Line, Bar } from "react-chartjs-2";
-import Icon from "../Shared/Icon";
-import ExpandIcon from "../../styles/svg/expand.svg";
-import CollapseIcon from "../../styles/svg/collapse.svg";
 
 const DeveloperDashboard = ({ currentDashboard, monthlyCashflow }) => {
   const [showPreFinanceCashflow, setShowPreFinanceCashflow] = useState(true);
@@ -32,58 +30,62 @@ const DeveloperDashboard = ({ currentDashboard, monthlyCashflow }) => {
   const total = tableData?.totalCashflow;
   return (
     <>
-      <h2 className="f16 bold mt16 mb16">Pre Funding Metrics</h2>
-      <div className="dash-row flex-row">
-        <table className="dashboard-table r bs-3 bg-1 p20 mb16 o-hidden">
+      <h2 className="my-2 text-l font-semibold">Pre Funding Metrics</h2>
+      <div className="flex flex-col md:flex-row mb-2">
+        <table className="flex-1 mb-1 md:my-0 md:mr-1 text-left shadow-xl rounded-2xl bg-white overflow-hidden">
           <thead>
-            <tr>
-              <th>Total</th>
+            <tr className="border-b border-gray-200 h-8 text-sm">
+              <th className="font-semibold p-3">Total</th>
               <th></th>
             </tr>
           </thead>
-          <tbody>
-            <tr>
-              <td className="text-s">Revenue:</td>
-              <td>{cashflowFormatter(total?.totalRevenue)}</td>
+          <tbody className="text-sm">
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">Revenue</td>
+              <td className="p-3">{cashflowFormatter(total?.totalRevenue)}</td>
             </tr>
-            <tr>
-              <td className="text-s">Costs:</td>
-              <td>{cashflowFormatter(total?.totalCostsPreFinance)}</td>
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">Costs</td>
+              <td className="p-3">
+                {cashflowFormatter(total?.totalCostsPreFinance)}
+              </td>
             </tr>
-            <tr>
-              <td className="text-s">Profit:</td>
-              <td>{cashflowFormatter(total?.preFinanceCashflow)}</td>
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">Profit</td>
+              <td className="p-3">
+                {cashflowFormatter(total?.preFinanceCashflow)}
+              </td>
             </tr>
           </tbody>
         </table>
-        <table className="dashboard-table r bs-3 bg-1 p20 mb16 o-hidden">
+        <table className="flex-1 mb-1 md:my-0 md:mr-1 text-left shadow-xl rounded-2xl bg-white overflow-hidden">
           <thead>
-            <tr>
-              <th>Per Dwelling</th>
+            <tr className="border-b border-gray-200 h-8 text-sm">
+              <th className="font-semibold p-3">Per Dwelling</th>
               <th></th>
             </tr>
           </thead>
-          <tbody>
-            <tr>
-              <td className="text-s">Revenue:</td>
-              <td>
+          <tbody className="text-sm">
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">Revenue</td>
+              <td className="p-3">
                 {cashflowFormatter(
                   total?.totalRevenue / currentDashboard.assumptions.dwellings
                 )}
               </td>
             </tr>
-            <tr>
-              <td className="text-s">Costs:</td>
-              <td>
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">Costs</td>
+              <td className="p-3">
                 {cashflowFormatter(
                   total?.totalCostsPreFinance /
                     currentDashboard.assumptions.dwellings
                 )}
               </td>
             </tr>
-            <tr>
-              <td className="text-s">Profit:</td>
-              <td>
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">Profit</td>
+              <td className="p-3">
                 {cashflowFormatter(
                   total?.preFinanceCashflow /
                     currentDashboard.assumptions.dwellings
@@ -92,25 +94,25 @@ const DeveloperDashboard = ({ currentDashboard, monthlyCashflow }) => {
             </tr>
           </tbody>
         </table>
-        <table className="dashboard-table r bs-3 bg-1 p20 mb16 o-hidden">
+        <table className="flex-1 mb-1 md:my-0 md:mr-1 text-left shadow-xl rounded-2xl bg-white overflow-hidden">
           <thead>
-            <tr>
-              <th>Unlevered Metrics</th>
+            <tr className="border-b border-gray-200 h-8 text-sm">
+              <th className="font-semibold p-3">Unlevered Metrics</th>
               <th></th>
             </tr>
           </thead>
-          <tbody>
-            <tr>
-              <td className="text-s">IRR:</td>
-              <td>
+          <tbody className="text-sm">
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">IRR</td>
+              <td className="p-3 text-gray-500">
                 {percentageFormatter.format(
                   IRRCalculation(monthlyCashflow).preFinance
                 )}
               </td>
             </tr>
-            <tr>
-              <td className="text-s">Margin On Cost:</td>
-              <td>
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">Margin On Cost</td>
+              <td className="p-3 text-gray-500">
                 {percentageFormatter.format(
                   developerMOCCalculation(monthlyCashflow).preFinance
                 )}
@@ -119,133 +121,153 @@ const DeveloperDashboard = ({ currentDashboard, monthlyCashflow }) => {
           </tbody>
         </table>
       </div>
-      <div className="dash-chart r bs-3 bg-1 p20 mb16">
-        <h3 className="f16 mb16">Cumulative Position Before Funding</h3>
+      <div className="shadow-xl rounded-2xl bg-white p-3 mb-2">
+        <h3 className="text-l mb-3">Cumulative Cash Position Before Funding</h3>
         <Line
           data={cumulativeChart.data.preFinance}
           options={cumulativeChart.options}
         />
       </div>
-      <div className="dash-cashflow r bs-3 bg-1 p20 mb16">
-        <div className="relative">
-          <div className="flex-row align-c mb16 mr100">
-            <h3 className="f16 mt-md-48">Cashflow Summary Before Funding</h3>
+      <div className="shadow-xl rounded-2xl bg-white p-3">
+        <div
+          className={`flex items-center justify-between text-l ${
+            showPreFinanceCashflow ? "mb-3" : ""
+          }`}
+        >
+          <div className="flex mr-3">
+            <h3 className="text-l mr-3">Annual Cashflow Before Funding</h3>
             <Tooltip message={preFinanceMessage} extraClass="mt8 mt-md-48" />
           </div>
-          <button
-            type="button"
-            className="absolute form-button-p bs-3 font-white pt4 pb4 flex-row align-c justify-c"
-            onClick={() => setShowPreFinanceCashflow(!showPreFinanceCashflow)}
-          >
-            <Icon
-              size={"20px"}
-              url={showPreFinanceCashflow ? CollapseIcon : ExpandIcon}
-              color={"white"}
-              hover={false}
-              active={false}
-            />
-            <span className="ml8">
-              {showPreFinanceCashflow ? "Hide" : "Show"}
-            </span>
-          </button>
+          <Button
+            label={showPreFinanceCashflow ? "Hide" : "Show"}
+            type={"button"}
+            options={{
+              styleType: "primary",
+              buttonClass: "h-10 px-2",
+              icon: showPreFinanceCashflow ? "collapse" : "expand",
+              iconClass: "h-8 w-8 mr-2",
+              onClick: () => setShowPreFinanceCashflow(!showPreFinanceCashflow),
+            }}
+          />
         </div>
         {showPreFinanceCashflow && (
-          <div className="o-x-auto">
-            <table className="dashboard-table w100 bg-1 p20 mb16 o-hidden">
+          <div className="overflow-x-auto">
+            <table className="w-full text-left">
               <thead>
-                <tr>
-                  <th>Year</th>
-                  <th>Acquisition Costs</th>
-                  <th>TDC</th>
-                  <th>NOI</th>
-                  <th>Net Sale</th>
-                  <th>Total Income</th>
-                  <th>Total Costs</th>
-                  <th>Net Annual Cashflow</th>
+                <tr className="border-b border-gray-200 h-8 text-sm">
+                  <th className="font-semibold p-3">Year</th>
+                  <th className="font-semibold p-3">Acquisition Costs</th>
+                  <th className="font-semibold p-3">TDC</th>
+                  <th className="font-semibold p-3">NOI</th>
+                  <th className="font-semibold p-3">Net Sale</th>
+                  <th className="font-semibold p-3">Total Income</th>
+                  <th className="font-semibold p-3">Total Costs</th>
+                  <th className="font-semibold p-3">Net Annual Cashflow</th>
                 </tr>
               </thead>
-              <tbody>
+              <tbody className="text-sm">
                 {tableData?.annualCashflow.map((c) => (
-                  <tr key={c.year}>
-                    <td>{c.year}</td>
-                    <td>{cashflowFormatter(c.acquisitionCosts)}</td>
-                    <td>{cashflowFormatter(c.TDC)}</td>
-                    <td>{cashflowFormatter(c.NOI)}</td>
-                    <td>{cashflowFormatter(c.netSale)}</td>
-                    <td>{cashflowFormatter(c.totalIncome)}</td>
-                    <td>{cashflowFormatter(c.preFinanceTotalCost)}</td>
-                    <td>{cashflowFormatter(c.preFinanceCashflow)}</td>
+                  <tr
+                    className="border-b border-gray-200 hover:bg-gray-50"
+                    key={c.year}
+                  >
+                    <td className="p-3">{c.year}</td>
+                    <td className="p-3">
+                      {cashflowFormatter(c.acquisitionCosts)}
+                    </td>
+                    <td className="p-3">{cashflowFormatter(c.TDC)}</td>
+                    <td className="p-3">{cashflowFormatter(c.NOI)}</td>
+                    <td className="p-3">{cashflowFormatter(c.netSale)}</td>
+                    <td className="p-3">{cashflowFormatter(c.totalIncome)}</td>
+                    <td className="p-3">
+                      {cashflowFormatter(c.preFinanceTotalCost)}
+                    </td>
+                    <td className="p-3">
+                      {cashflowFormatter(c.preFinanceCashflow)}
+                    </td>
                   </tr>
                 ))}
               </tbody>
-              <tfoot>
-                <tr>
-                  <th>Total</th>
-                  <td>{cashflowFormatter(total?.acquisitionCosts)}</td>
-                  <td>{cashflowFormatter(total?.TDC)}</td>
-                  <td>{cashflowFormatter(total?.NOI)}</td>
-                  <td>{cashflowFormatter(total?.netSale)}</td>
-                  <td>{cashflowFormatter(total?.totalIncome)}</td>
-                  <td>{cashflowFormatter(total?.preFinanceTotalCost)}</td>
-                  <td>{cashflowFormatter(total?.preFinanceCashflow)}</td>
+              <tfoot className="text-sm">
+                <tr className="font-semibold border-b border-gray-200 hover:bg-gray-50">
+                  <td className="p-3">Total</td>
+                  <td className="p-3">
+                    {cashflowFormatter(total?.acquisitionCosts)}
+                  </td>
+                  <td className="p-3">{cashflowFormatter(total?.TDC)}</td>
+                  <td className="p-3">{cashflowFormatter(total?.NOI)}</td>
+                  <td className="p-3">{cashflowFormatter(total?.netSale)}</td>
+                  <td className="p-3">
+                    {cashflowFormatter(total?.totalIncome)}
+                  </td>
+                  <td className="p-3">
+                    {cashflowFormatter(total?.preFinanceTotalCost)}
+                  </td>
+                  <td className="p-3">
+                    {cashflowFormatter(total?.preFinanceCashflow)}
+                  </td>
                 </tr>
               </tfoot>
             </table>
           </div>
         )}
       </div>
-      <h2 className="f16 bold mt16 mb16">Post Funding Metrics</h2>
-      <div className="dash-row flex-row">
-        <table className="dashboard-table r bs-3 bg-1 p20 mb16 o-hidden">
+      <h2 className="my-2 text-l font-semibold">Post Funding Metrics</h2>
+      <div className="flex flex-col md:flex-row mb-2">
+        <table className="flex-1 mb-1 md:my-0 md:mr-1 text-left shadow-xl rounded-2xl bg-white overflow-hidden">
           <thead>
-            <tr>
-              <th>Total</th>
+            <tr className="border-b border-gray-200 h-8 text-sm">
+              <th className="font-semibold p-3">Total</th>
               <th></th>
             </tr>
           </thead>
-          <tbody>
-            <tr>
-              <td className="text-s">Revenue:</td>
-              <td>{cashflowFormatter(total?.totalRevenue)}</td>
+          <tbody className="text-sm">
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">Revenue</td>
+              <td className="p-3">{cashflowFormatter(total?.totalRevenue)}</td>
             </tr>
-            <tr>
-              <td className="text-s">Costs:</td>
-              <td>{cashflowFormatter(total?.totalCostsPostFinance)}</td>
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">Costs</td>
+              <td className="p-3">
+                {cashflowFormatter(total?.totalCostsPostFinance)}
+              </td>
             </tr>
-            <tr>
-              <td className="text-s">Profit:</td>
-              <td>{cashflowFormatter(total?.postFinanceCashflow)}</td>
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">Profit</td>
+              <td className="p-3">
+                {cashflowFormatter(total?.postFinanceCashflow)}
+              </td>
             </tr>
           </tbody>
         </table>
-        <table className="dashboard-table r bs-3 bg-1 p20 mb16 o-hidden">
+        <table className="flex-1 mb-1 md:my-0 md:mr-1 text-left shadow-xl rounded-2xl bg-white overflow-hidden">
           <thead>
-            <tr>
-              <th>Per Dwelling</th>
+            <tr className="border-b border-gray-200 h-8 text-sm">
+              <th className="font-semibold p-3">Per Dwelling</th>
               <th></th>
             </tr>
           </thead>
-          <tbody>
-            <tr>
-              <td className="text-s">Revenue:</td>
-              <td>
+          <tbody className="text-sm">
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">Revenue</td>
+              <td className="p-3">
                 {cashflowFormatter(
                   total?.totalRevenue / currentDashboard.assumptions.dwellings
                 )}
               </td>
             </tr>
-            <tr>
-              <td className="text-s">Costs:</td>
-              <td>
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">Costs</td>
+              <td className="p-3">
                 {cashflowFormatter(
                   total?.totalCostsPostFinance /
                     currentDashboard.assumptions.dwellings
                 )}
               </td>
             </tr>
-            <tr>
-              <td className="text-s">Profit:</td>
-              <td>
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">Profit</td>
+              <td className="p-3">
                 {cashflowFormatter(
                   total?.postFinanceCashflow /
                     currentDashboard.assumptions.dwellings
@@ -254,25 +276,25 @@ const DeveloperDashboard = ({ currentDashboard, monthlyCashflow }) => {
             </tr>
           </tbody>
         </table>
-        <table className="dashboard-table r bs-3 bg-1 p20 mb16 o-hidden">
+        <table className="flex-1 mb-1 md:my-0 md:mr-1 text-left shadow-xl rounded-2xl bg-white overflow-hidden">
           <thead>
-            <tr>
-              <th>Levered Metrics</th>
+            <tr className="border-b border-gray-200 h-8 text-sm">
+              <th className="font-semibold p-3">Levered Metrics</th>
               <th></th>
             </tr>
           </thead>
-          <tbody>
-            <tr>
-              <td className="text-s">IRR:</td>
-              <td>
+          <tbody className="text-sm">
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">IRR</td>
+              <td className="p-3 text-gray-500">
                 {percentageFormatter.format(
                   IRRCalculation(monthlyCashflow).postFinance
                 )}
               </td>
             </tr>
-            <tr>
-              <td className="text-s">Margin On Cost:</td>
-              <td>
+            <tr className="border-b border-gray-200 h-12 hover:bg-gray-50">
+              <td className="p-3">Margin On Cost</td>
+              <td className="p-3 text-gray-500">
                 {percentageFormatter.format(
                   developerMOCCalculation(monthlyCashflow).postFinance
                 )}
@@ -281,89 +303,108 @@ const DeveloperDashboard = ({ currentDashboard, monthlyCashflow }) => {
           </tbody>
         </table>
       </div>
-      <div className="dash-chart r bs-3 bg-1 p20 mb16">
-        <h3 className="f16 mb16">Cumulative Position After Funding</h3>
+      <div className="shadow-xl rounded-2xl bg-white p-3 mb-2">
+        <h3 className="text-l mb-3">Cumulative Position After Funding</h3>
         <Line
           data={cumulativeChart.data.postFinance}
           options={cumulativeChart.options}
         />
       </div>
-      <div className="dash-chart r bs-3 bg-1 p20 mb16">
-        <h3 className="f16 mb16">Cumulative Funding Uses</h3>
+      <div className="shadow-xl rounded-2xl bg-white p-3 mb-2">
+        <h3 className="text-l mb-3">Cumulative Funding Uses</h3>
         <Line data={fundingChart.data} options={fundingChart.options} />
       </div>
-      <div className="dash-chart r bs-3 bg-1 p20 mb16">
-        <h3 className="f16 mb16">Annual Cashflow</h3>
+      <div className="shadow-xl rounded-2xl bg-white p-3 mb-2">
+        <h3 className="text-l mb-3">Annual Cashflow</h3>
         <Bar data={annualChart.data} options={annualChart.options} />
       </div>
-      <div className="dash-cashflow r bs-3 bg-1 p20 mb16">
-        <div className="relative">
-          <div className="flex-row align-c mb16 mr100">
-            <h3 className="f16 mt-md-48">Cashflow Summary After Funding</h3>
+      <div className="shadow-xl rounded-2xl bg-white p-3 mb-2">
+        <div
+          className={`flex items-center justify-between text-l ${
+            showPostFinanceCashflow ? "mb-3" : ""
+          }`}
+        >
+          <div className="flex mr-3">
+            <h3 className="text-l mr-3">Annual Cashflow After Funding</h3>
             <Tooltip message={postFinanceMessage} extraClass="mt8 mt-md-48" />
           </div>
-          <button
-            type="button"
-            className="absolute form-button-p bs-3 font-white pt4 pb4 flex-row align-c justify-c"
-            onClick={() => setShowPostFinanceCashflow(!showPostFinanceCashflow)}
-          >
-            <Icon
-              size={"20px"}
-              url={showPostFinanceCashflow ? CollapseIcon : ExpandIcon}
-              color={"white"}
-              hover={false}
-              active={false}
-            />
-            <span className="ml8">
-              {showPostFinanceCashflow ? "Hide" : "Show"}
-            </span>
-          </button>
+          <Button
+            label={showPostFinanceCashflow ? "Hide" : "Show"}
+            type={"button"}
+            options={{
+              styleType: "primary",
+              buttonClass: "h-10 px-2",
+              icon: showPostFinanceCashflow ? "collapse" : "expand",
+              iconClass: "h-8 w-8 mr-2",
+              onClick: () =>
+                setShowPostFinanceCashflow(!showPostFinanceCashflow),
+            }}
+          />
         </div>
         {showPostFinanceCashflow && (
-          <div className="o-x-auto">
-            <table className="dashboard-table w100 bg-1 p20 mb16 o-hidden">
+          <div className="overflow-x-auto">
+            <table className="w-full text-left">
               <thead>
-                <tr>
-                  <th>Year</th>
-                  <th>Acquisition Costs</th>
-                  <th>TDC</th>
-                  <th>NOI</th>
-                  <th>Net Sale</th>
-                  <th>Funding Costs</th>
-                  <th>Total Income</th>
-                  <th>Total Costs</th>
-                  <th>Debt Source</th>
-                  <th>Net Annual Cashflow</th>
+                <tr className="border-b border-gray-200 h-8 text-sm">
+                  <th className="font-semibold p-3">Year</th>
+                  <th className="font-semibold p-3">Acquisition Costs</th>
+                  <th className="font-semibold p-3">TDC</th>
+                  <th className="font-semibold p-3">NOI</th>
+                  <th className="font-semibold p-3">Net Sale</th>
+                  <th className="font-semibold p-3">Funding Costs</th>
+                  <th className="font-semibold p-3">Total Income</th>
+                  <th className="font-semibold p-3">Total Costs</th>
+                  <th className="font-semibold p-3">Debt Source</th>
+                  <th className="font-semibold p-3">Net Annual Cashflow</th>
                 </tr>
               </thead>
-              <tbody>
+              <tbody className="text-sm">
                 {tableData?.annualCashflow.map((c) => (
-                  <tr key={c.year}>
-                    <td>{c.year}</td>
-                    <td>{cashflowFormatter(c.acquisitionCosts)}</td>
-                    <td>{cashflowFormatter(c.TDC)}</td>
-                    <td>{cashflowFormatter(c.NOI)}</td>
-                    <td>{cashflowFormatter(c.netSale)}</td>
-                    <td>{cashflowFormatter(c.loanCosts)}</td>
-                    <td>{cashflowFormatter(c.totalIncome)}</td>
-                    <td>{cashflowFormatter(c.postFinanceTotalCost)}</td>
-                    <td>{cashflowFormatter(c.debtSource)}</td>
-                    <td>{cashflowFormatter(c.postFinanceCashflow)}</td>
+                  <tr
+                    className="border-b border-gray-200 hover:bg-gray-50"
+                    key={c.year}
+                  >
+                    <td className="p-3">{c.year}</td>
+                    <td className="p-3">
+                      {cashflowFormatter(c.acquisitionCosts)}
+                    </td>
+                    <td className="p-3">{cashflowFormatter(c.TDC)}</td>
+                    <td className="p-3">{cashflowFormatter(c.NOI)}</td>
+                    <td className="p-3">{cashflowFormatter(c.netSale)}</td>
+                    <td className="p-3">{cashflowFormatter(c.loanCosts)}</td>
+                    <td className="p-3">{cashflowFormatter(c.totalIncome)}</td>
+                    <td className="p-3">
+                      {cashflowFormatter(c.postFinanceTotalCost)}
+                    </td>
+                    <td className="p-3">{cashflowFormatter(c.debtSource)}</td>
+                    <td className="p-3">
+                      {cashflowFormatter(c.postFinanceCashflow)}
+                    </td>
                   </tr>
                 ))}
               </tbody>
-              <tfoot>
-                <tr>
-                  <th>Total</th>
-                  <td>{cashflowFormatter(total?.acquisitionCosts)}</td>
-                  <td>{cashflowFormatter(total?.TDC)}</td>
-                  <td>{cashflowFormatter(total?.NOI)}</td>
-                  <td>{cashflowFormatter(total?.netSale)}</td>
-                  <td>{cashflowFormatter(total?.loanCosts)}</td>
-                  <td>{cashflowFormatter(total?.totalIncome)}</td>
-                  <td>{cashflowFormatter(total?.postFinanceTotalCost)}</td>
-                  <td>{cashflowFormatter(total?.debtSource)}</td>
-                  <td>{cashflowFormatter(total?.postFinanceCashflow)}</td>
+              <tfoot className="text-sm">
+                <tr className="font-semibold border-b border-gray-200 hover:bg-gray-50">
+                  <td className="p-3">Total</td>
+                  <td className="p-3">
+                    {cashflowFormatter(total?.acquisitionCosts)}
+                  </td>
+                  <td className="p-3">{cashflowFormatter(total?.TDC)}</td>
+                  <td className="p-3">{cashflowFormatter(total?.NOI)}</td>
+                  <td className="p-3">{cashflowFormatter(total?.netSale)}</td>
+                  <td className="p-3">{cashflowFormatter(total?.loanCosts)}</td>
+                  <td className="p-3">
+                    {cashflowFormatter(total?.totalIncome)}
+                  </td>
+                  <td className="p-3">
+                    {cashflowFormatter(total?.postFinanceTotalCost)}
+                  </td>
+                  <td className="p-3">
+                    {cashflowFormatter(total?.debtSource)}
+                  </td>
+                  <td className="p-3">
+                    {cashflowFormatter(total?.postFinanceCashflow)}
+                  </td>
                 </tr>
               </tfoot>
             </table>
