+++ client/src/components/Shared/Overlay.js
@@ -1,10 +1,12 @@
 import React from "react";
 import { OverlayOutsideAlerter } from "../../utils/hooks";
 
-const Overlay = () => {
+const Overlay = (props) => {
   return (
     <OverlayOutsideAlerter>
-      <div className="overlay fade-in" />
+      <div className="w-screen h-screen animate-fade-in fixed left-0 top-0 z-40 bg-gray-800 bg-opacity-70 flex justify-center items-center">
+        {props.children}
+      </div>
     </OverlayOutsideAlerter>
   );
 };
